client/server common IPC
The client executable is now tiny and will be removed: the Main.hs
program will be extended so it can express the current client.
