+++ byron-proxy/byron-proxy.cabal
@@ -66,7 +66,8 @@ library
 
 executable byron-proxy
   main-is:             Main.hs
-  other-modules:       DB
+  other-modules:       DB,
+                       IPC
   -- other-extensions:
   build-depends:       base,
                        async,
@@ -109,7 +110,8 @@ executable byron-proxy
 
 executable byron-proxy-client-demo
   main-is:             Client.hs
-  other-modules:       DB
+  other-modules:       DB,
+                       IPC
   -- other-extensions:
   build-depends:       base,
                        async,
@@ -123,10 +125,15 @@ executable byron-proxy-client-demo
                        containers,
                        contra-tracer,
                        directory,
+                       exceptions,
+                       io-sim-classes,
                        network,
                        ouroboros-consensus,
                        ouroboros-network,
+                       resourcet,
                        serialise,
+                       stm,
+                       transformers,
                        typed-protocols
   hs-source-dirs:      src/exec
   default-language:    Haskell2010
